# api/openapi.yaml
openapi: 3.0.3
info:
  title: WhereToLive HTTP API
  version: 0.1.0
  description: Authoritative HTTP contract for the WhereToLive modular monolith.
servers:
  - url: /api
paths:
...
  /platform/settings/dictionaries:
    get:
      operationId: listPlatformDictionaries
      summary: List global dictionaries as a platform administrator
      tags: [platform-settings]
      security: [{ platformBearerAuth: [] }]
      responses:
        "200":
          description: Global dictionaries
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Dictionary" } }
        "401": { $ref: "#/components/responses/AuthenticationFailed" }
  /platform/settings/dictionaries/{dictionaryCode}:
    put:
      operationId: replacePlatformDictionary
      summary: Replace a complete global dictionary
      tags: [platform-settings]
      security: [{ platformBearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/DictionaryCode" }
        - { $ref: "#/components/parameters/CsrfToken" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReplacePlatformDictionaryRequest" }
      responses:
        "200":
          description: Global dictionary desired state stored
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Dictionary" }
        "400": { description: Invalid request }
        "401": { $ref: "#/components/responses/AuthenticationFailed" }
        "403": { $ref: "#/components/responses/CsrfFailed" }
        "409": { description: Version conflict }
  /platform/settings/configurations:
    get:
      operationId: listPlatformConfigurations
      summary: List global non-secret configuration definitions and defaults
      tags: [platform-settings]
      security: [{ platformBearerAuth: [] }]
      responses:
        "200":
          description: Global configurations
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Configuration" } }
        "401": { $ref: "#/components/responses/AuthenticationFailed" }
  /platform/settings/configurations/{configurationKey}:
    put:
      operationId: putPlatformConfiguration
      summary: Create or update a global non-secret configuration
      tags: [platform-settings]
      security: [{ platformBearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/ConfigurationKey" }
        - { $ref: "#/components/parameters/CsrfToken" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PutPlatformConfigurationRequest" }
      responses:
        "200":
          description: Global configuration stored
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Configuration" }
        "400": { description: Invalid request }
        "401": { $ref: "#/components/responses/AuthenticationFailed" }
        "403": { $ref: "#/components/responses/CsrfFailed" }
        "409": { description: Version conflict }
...
components:
  securitySchemes:
    consumerBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    platformBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 32
        maxLength: 256
...
    DictionaryCode:
      name: dictionaryCode
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 64 }
    ConfigurationKey:
      name: configurationKey
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 64 }
...
  responses:
...
    AuthenticationFailed:
      description: Tenant or credentials are invalid, or the session is unavailable
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    CsrfFailed:
      description: CSRF validation failed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
...
    SettingSource:
      type: string
      enum: [global, tenant]
    ConfigurationValueType:
      type: string
      enum: [string, boolean, integer, json]
    DictionaryItem:
      type: object
      additionalProperties: false
      required: [code, label, sortOrder, enabled]
      properties:
        code: { type: string, minLength: 1, maxLength: 64 }
        label: { type: string, minLength: 1, maxLength: 128 }
        sortOrder: { type: integer }
        enabled: { type: boolean }
    Dictionary:
      type: object
      additionalProperties: false
      required: [code, name, source, version, items]
      properties:
        code: { type: string }
        name: { type: string }
        source: { $ref: "#/components/schemas/SettingSource" }
        version: { type: integer, format: int64, minimum: 1 }
        items:
          type: array
          items: { $ref: "#/components/schemas/DictionaryItem" }
...
    ReplacePlatformDictionaryRequest:
      type: object
      additionalProperties: false
      required: [name, expectedVersion, items, reason]
      properties:
        name: { type: string, minLength: 1, maxLength: 128 }
        expectedVersion: { type: integer, format: int64, minimum: 0 }
        items:
          type: array
          maxItems: 500
          items: { $ref: "#/components/schemas/DictionaryItem" }
        reason: { type: string, minLength: 1, maxLength: 500 }
    Configuration:
      type: object
      additionalProperties: false
      required: [key, name, valueType, tenantOverridable, source, version, value]
      properties:
        key: { type: string }
        name: { type: string }
        valueType: { $ref: "#/components/schemas/ConfigurationValueType" }
        tenantOverridable: { type: boolean }
        source: { $ref: "#/components/schemas/SettingSource" }
        version: { type: integer, format: int64, minimum: 1 }
        value: {}
...
    PutPlatformConfigurationRequest:
      type: object
      additionalProperties: false
      required: [name, valueType, tenantOverridable, expectedVersion, value, reason]
      properties:
        name: { type: string, minLength: 1, maxLength: 128 }
        valueType: { $ref: "#/components/schemas/ConfigurationValueType" }
        tenantOverridable: { type: boolean }
        expectedVersion: { type: integer, format: int64, minimum: 0 }
        value: {}
        reason: { type: string, minLength: 1, maxLength: 500 }
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        traceId:
          type: string